modularizacion companyRegister y studentRegister -gdd
