+++ src/App.jsx
@@ -2,11 +2,11 @@ import { useState } from "react";
 import reactLogo from "./assets/react.svg";
 import viteLogo from "/vite.svg";
 import "./App.css";
-import RegisterStudent from "./components/registerStudent";
+import StudentRegister from "./views/studentRegister/studentRegister";
 import { Route, Routes ,Link} from "react-router-dom";
 import Home from "./components/Home";
-import CompanyRegister from "./views/companyRegister";
-//import RegisterCompany from "./components/registerCompany";
+import CompanyRegister from "./views/companyRegister/companyRegister";
+
 
 
 function App() {
@@ -43,7 +43,7 @@ function App() {
 				{/* <RegisterStudent/> */}
                 <Routes>
                     <Route exact path='/Trends_app_MVP' element={<Home/>} />
-                    <Route exact path='/registerStudent' element={<RegisterStudent/>} />
+                    <Route exact path='/studentRegister' element={<StudentRegister/>} />
                     <Route exact path='/companyRegister' element={<CompanyRegister/>} />
                 </Routes>
 			</div>
